requirements.txt
nextcord~=2.0.0a5
asyncio
Pillow>=8.3.2
pytesseract~=0.3.8
requests>=2.22.0
algoliasearch>=2.6.0
packaging
jellyfish
python-logstash-async
google-cloud-dialogflow
sqlobject~=3.9.1
psycopg2-binary~=2.9.0
aiohttp
python-dotenv
jsonpickle